Replace [po_directory] by [po4a_langs] and [po4a_paths] in cfg files
This is not recommended practice, but in our case it helps to
provide a framework which can be used by translation teams.

Po4a >= 0.41 is required to generate POT files when $(langs) is
empty.  But there is no extra requirement when generating POT
files for translated manual pages, ie. when $(langs) is not empty.